+++ lib/helpers/column-set.js
@@ -13,7 +13,7 @@ const npm = {
     os: require('os'),
     utils: require('../utils'),
     formatting: require('../formatting'),
-    TableName: require('./tableName'),
+    TableName: require('./table-name'),
     Column: require('./column')
 };
