+++ lib/helpers/index.js
@@ -14,8 +14,8 @@ const npm = {
     update: require('./methods/update'),
     values: require('./methods/values'),
     sets: require('./methods/sets'),
-    TableName: require('./tableName'),
-    ColumnSet: require('./columnSet'),
+    TableName: require('./table-name'),
+    ColumnSet: require('./column-set'),
     Column: require('./column')
 };
