+++ lib/helpers/methods/insert.js
@@ -8,8 +8,8 @@
  */
 
 const npm = {
-    TableName: require('../tableName'),
-    ColumnSet: require('../columnSet'),
+    TableName: require('../table-name'),
+    ColumnSet: require('../column-set'),
     formatting: require('../../formatting'),
     utils: require('../../utils')
 };
