+++ lib/helpers/methods/sets.js
@@ -8,7 +8,7 @@
  */
 
 const npm = {
-    ColumnSet: require('../columnSet'),
+    ColumnSet: require('../column-set'),
     format: require('../../formatting').as.format,
     utils: require('../../utils')
 };
