+++ lib/helpers/methods/update.js
@@ -10,8 +10,8 @@
 const {assertOptions} = require('assert-options');
 
 const npm = {
-    TableName: require('../tableName'),
-    ColumnSet: require('../columnSet'),
+    TableName: require('../table-name'),
+    ColumnSet: require('../column-set'),
     formatting: require('../../formatting'),
     utils: require('../../utils')
 };
