+++ lib/helpers/methods/values.js
@@ -8,7 +8,7 @@
  */
 
 const npm = {
-    ColumnSet: require('../columnSet'),
+    ColumnSet: require('../column-set'),
     formatting: require('../../formatting'),
     utils: require('../../utils')
 };
