+++ lib/helpers/table-name.js
