+++ lib/main.js
@@ -25,7 +25,7 @@ const npm = {
     errors: require('./errors'),
     utils: require('./utils'),
     pubUtils: require('./utils/public'),
-    mode: require('./txMode'),
+    mode: require('./tx-mode'),
     package: require('../package.json'),
     text: require('./text')
 };
