+++ lib/task.js
@@ -12,7 +12,7 @@ const {Events} = require('./events');
 const npm = {
     spex: require('spex'),
     utils: require('./utils'),
-    mode: require('./txMode'),
+    mode: require('./tx-mode'),
     query: require('./query'),
     async: require('./async'),
     text: require('./text')
