+++ lib/tx-mode.js
